requirements.txt
pandas
streamlit
chromadb
langchain-core
langchain-openai
langchain-community
python-dotenv
streamlit